✨ Convert Client to TS Project
