+++ package.json
@@ -10,8 +10,8 @@
     "format": "prettier --write .",
     "lint": "eslint \"src/**/*.ts\"",
     "test": "echo \"Error: no test specified\" && exit 0",
-    "build-client": "webpack",
-    "build-server": "npm run format && npm run lint && tsc",
+    "build-client": "rimraf dist && env-cmd webpack",
+    "build-server": "rimraf build && npm run format && npm run lint && env-cmd tsc",
     "build": "npm run build-client && npm run build-server",
     "start": "node --no-warnings build/index.js",
     "dev": "env-cmd npm start",
@@ -41,6 +41,8 @@
     "@types/node-fetch": "^2.5.5",
     "@types/papaparse": "^5.0.3",
     "@types/progress": "^2.0.3",
+    "@types/react": "^16.9.34",
+    "@types/react-dom": "^16.9.6",
     "@types/sharp": "^0.24.0",
     "@types/uuid": "^7.0.2",
     "@typescript-eslint/eslint-plugin": "^2.26.0",
@@ -56,7 +58,10 @@
     "eslint-plugin-react": "^7.19.0",
     "html-webpack-plugin": "^4.0.4",
     "prettier": "^2.0.2",
+    "rimraf": "^3.0.2",
+    "source-map-loader": "^0.2.4",
     "style-loader": "^1.1.3",
+    "ts-loader": "^6.2.2",
     "typescript": "^3.8.3",
     "url-loader": "^4.0.0",
     "webpack": "^4.42.1",
