+++ public/index.html
@@ -7,5 +7,15 @@
   </head>
   <body>
     <div id="root"></div>
+    <!-- <script
+      crossorigin
+      src="https://unpkg.com/react@16/umd/react.development.js"
+    ></script>
+    <script
+      crossorigin
+      src="https://unpkg.com/react-dom@16/umd/react-dom.development.js"
+    ></script> -->
+    <script src="/scripts/react.js"></script>
+    <script src="/scripts/react-dom.js"></script>
   </body>
 </html>
