+++ typings/custom/import-png.d.ts
@@ -1,0 +1,5 @@
+declare module '*.png' {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const value: any;
+  export default value;
+}
